@@ -13,8 +13,8 @@
     "formik": "^2.4.2",
     "modern-normalize": "^2.0.0",
     "prop-types": "^15.8.1",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
     "react-scripts": "5.0.1",
     "react-toastify": "^9.1.3",
